Don't ignore values.class when string type
The lodash.union function ignores arguments that are string type, therefore `values.class` of type string are not added to the `classes` array, and don't end up in the resulting HTML string.

First look at the lodash union function instantiation:

https://github.com/lodash/lodash/blob/master/dist/lodash.js#L7579

The pertinent argument (last one in this version) is `true`. Now look at the `baseFlatten` function, where a string is handled when union'izing:

https://github.com/lodash/lodash/blob/master/dist/lodash.js#L2613-2614

...the corresponding argument (isStrict) must be `false` in order for strings to be included in the resulting array. Therefore strings are ignored.

The reason for the `values.class && [values.class]` and not just `[values.class]` is because when `values.class` is undefined and wrapped in array brackets it still registers as an element on the array, so having only `[values.class]` will always produce the class attribute in the HTML string, but empty when `values.class` === undefined (i.e., <div class="">).